fixed conversion

+++ labor-neighbor-app/Client/src/Components/PostJob.js
@@ -148,7 +148,7 @@ export default class PostJob extends Component{
 		var data = {
   			"Title" : this.state.jobname, 
   			"NumLaborers" : numL,
-  			"Price" : this.state.wage,
+  			"Price" : (int) this.state.wage,
   			"City" : this.state.location,
   			"Name" : "Mike Williams",
   			"Description" : this.state.description
